bump BESS to 52501855
It's got DPDK 19.11.3, which has fancy new drivers

+++ generator/common.py
@@ -527,7 +527,7 @@ class Session(object):
                 tx_pipeline.modules[0].update(pps=pps_per_core)
             else:
                 self.__bess.update_tc_params(tc, resource='packet',
-                                             limit={'packet': long(pps_per_core)})
+                                             limit={'packet': int(pps_per_core)})
 
     def update_port_stats(self, now=None):
         if self.__last_stats is not None:
